Add script to allow for upgrade of Heroku deployments
Similar to the Heroku `deploy.sh`, it's handy to be able to run a script to easily upgrade a Postfacto deployment to the latest version. This change provides a simple script that updates the deployment in place.

Note that it doesn't do any management of environment variables that may have changed between deployments, so caveat emptor applies\!