Add color change to hover state of "thumbs up" and "edit pencil"
We find ourselves commonly clicking just outside of the click bounds because the cursor is always a "pointer" in the whole item. Having the color change specifically when you're over the actual click bounds should greatly reduce this pain.

The color that I picked matches the "video" and "menu" buttons found at the top right of the page and nicely fits with the theme as a whole.